refactor(github): move RepoAnalysisForm to new feature module
- Create new feature module for GitHub analysis
- Move RepoAnalysisForm from components to features/github/analysis
- Update import in GitHub.tsx to use new feature module
