+++ src/features/github/analysis/index.ts
@@ -1,0 +1,1 @@
+export { default as RepoAnalysisForm } from './RepoAnalysisForm';
